changed intervalInSeconds to double

+++ include/bubbleSort.h
@@ -1,6 +1,6 @@
 #ifndef BUBBLESORT_H
 #define BUBBLESORT_H
 
-void bubbleSort(int *array, int size, int intervalInSeconds);
+void bubbleSort(int *array, int size, double intervalInSeconds);
 
 #endif // BUBBLESORT_H
