+++ include/insertionSort.h
@@ -1,6 +1,6 @@
 #ifndef INSERTIONSORT_H
 #define INSERTIONSORT_H
 
-void insertionSort(int *array, int size, int intervalInSeconds);
+void insertionSort(int *array, int size, double intervalInSeconds);
 
 #endif // INSERTIONSORT_H
