+++ include/mergeSort.h
@@ -1,6 +1,6 @@
 #ifndef MERGESORT_H
 #define MERGESORT_H
 
-void mergeSort(int *array, int size, int intervalInSeconds);
+void mergeSort(int *array, int size, double intervalInSeconds);
 
 #endif // MERGESORT_H
